2 Версия ТЗ10

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -12,7 +12,7 @@ public interface FilmStorage {
 
     Film updateFIlm(Film film);
 
-    void deleteFilm(int id);
+    void deleteFilm(long id);
 
-    Film getFilmById(int id);
+    Film getFilmById(long id);
 }
